WIP Start adding new team members.
This doesn't work as expected, I suspect that the Team api call isn't
saving the player values properly. If you save the newly added player
the data gets stored somewhere but a new _id is not generated for that
player.

Obviously this screen still needs the ability to actually set/change the
players values.